end the blog app

+++ app/controllers/post_controller.ts
@@ -8,6 +8,8 @@ import {Marked} from 'marked'
 import {markedHighlight} from "marked-highlight"
 import hljs from 'highlight.js'
 import { unlink } from 'fs/promises'
+import { alterPost } from '#abilities/main'
+import PostPolicy from '#policies/post_policy'
 @inject()
 export default class PostController {
   constructor(private readonly fileUploaderService : FileUploaderService){}
@@ -17,7 +19,7 @@ export default class PostController {
    */
   async index({view,request}: HttpContext) {
     const page = request.input('page',1)
-    const limit = 3
+    const limit = 2
     const posts = await Post
                         .query()
                         .select('id','title','thumbnail','slug','user_id')
@@ -80,19 +82,27 @@ export default class PostController {
   /**
    * Edit individual record
    */
-  async edit({ params, view }: HttpContext) {
+  async edit({ params, view , bouncer, response, session}: HttpContext) {
     const {id} = params
     const post = await Post.findByOrFail('id',id)
+    if(await bouncer.with(PostPolicy).denies('alterPost', post)){
+      session.flash('error','Action interdite')
+      return response.redirect().back()
+    }
     return view.render('pages/post/edit', {post})
   }
 
   /**
    * Handle form submission for the edit action
    */
-  async update({ params, request,session, response }: HttpContext) {
+  async update({ params, request,session, response, bouncer }: HttpContext) {
     const {id} = params
     const {content, thumbnail,title} = await request.validateUsing(updatePostValidator)
     const post = await Post.findByOrFail('id',id)
+    if(await bouncer.with(PostPolicy).denies('alterPost', post)){
+      session.flash('error','Action interdite')
+      return response.redirect().back()
+    }
     const slug = post.title !== title && stringHelpers.slug(title)
     if(thumbnail){
       await unlink(`public/${post.thumbnail}`)
@@ -109,5 +119,16 @@ export default class PostController {
   /**
    * Delete record
    */
-  async destroy({ params }: HttpContext) {}
+  async destroy({ params, session, response, bouncer }: HttpContext) {
+    const {id} = params
+    const post = await Post.findByOrFail('id',id)
+    if(await bouncer.with(PostPolicy).denies('alterPost', post)){
+      session.flash('error','Action interdite')
+      return response.redirect().back()
+    }
+    await unlink(`public/${post.thumbnail}`)
+    await post.delete()
+    session.flash('success', "votre post a été bien supprimer")
+    return response.redirect().toRoute('home')
+  }
 }
